Introduce permission handler & unit test
